python-backend/requirements.txt
# Indie Game Platform - Backend Dependencies
# FastAPI und Web-Framework
fastapi==0.111.0
uvicorn[standard]==0.29.0

# Datenbank
sqlalchemy==2.0.29

# Authentifizierung
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9

# Validierung und Serialisierung
pydantic==2.7.1
email-validator==2.1.1

# File Upload
aiofiles==23.2.1

# Environment Variables
python-dotenv==1.0.1

# HTTP Client für Health Checks
requests==2.31.0

# PostgreSQL-Treiber (wird nur in Produktionsumgebungen installiert)
psycopg2-binary==2.9.9

# Development Dependencies (optional)
pytest==8.2.2
pytest-asyncio==0.23.7
black==24.4.2
flake8==7.1.0